HREF Parsing - Refactor Facebook
Add ability to parse through href. Add ability to launch multiple scripts from the entry point of the app. Add ability to inject Facebook dependencies when on the facebook website.
